+++ src/main/java/com/github/castorm/kafka/connect/http/response/timestamp/DateTimeFormatterTimestampParserConfig.java
@@ -47,7 +47,7 @@ public class DateTimeFormatterTimestampParserConfig extends AbstractConfig {
 
     public static ConfigDef config() {
         return new ConfigDef()
-                .define(ITEM_TIMESTAMP_PATTERN, STRING, "yyyy-MM-dd'T'HH:mm:ss.SSSX", LOW, "Timestamp format pattern")
+                .define(ITEM_TIMESTAMP_PATTERN, STRING, "yyyy-MM-dd'T'HH:mm:ss[.SSS]X", LOW, "Timestamp format pattern")
                 .define(ITEM_TIMESTAMP_ZONE, STRING, "UTC", LOW, "Timestamp ZoneId");
     }
 }