+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   `http.response.records.pointer` renamed to `http.response.list.pointer`
 -   `http.response.record.value.pointer` renamed to `http.response.record.pointer`
 -   Enable compound keys on `JacksonRecordParser`
+-   Change default value for `http.response.record.timestamp.parser.pattern` from `yyyy-MM-dd'T'HH:mm:ss.SSSX` to `yyyy-MM-dd'T'HH:mm:ss[.SSS]X`
 
 ## v0.6.3 (05/22/2020)
 -   Rename plugin folder name to match artifactId